tapnode: move group verifier generators out of tapgarden
GenGroupVerifier, GenGroupAnchorVerifier, and
GenRawGroupAnchorVerifier construct closures over
tapnode.GroupFetcher that feed proof.VerifierCtx; they sit naturally
alongside tapnode.GenHeaderVerifier rather than inside the minting
state machine. Verifier consumption is downstream of tapgarden's
end (a verifiable asset in the local store); the generators belong
to the proof-verifier code, not the minting one.

Move the three generators to tapnode along with
ErrGroupKeyUnknown, ErrGenesisNotGroupAnchor, and the LRU cache
helpers. Call sites are updated accordingly.
